chore: move letters to components
